[payroll-api] Bugfix in handlers
Resetting the handler object at the beginning of each HTTP request.
Since the same instance is used to populate response for the request,
and the fact that the instance is created inside main() and lives
throughout the duration of program, the responses keep piling up!

This means, if same GET request is sent twice, the second response will
have data from both first and second request (essentially duplicated
data) in the response. Lesson learnt! Preferably use separate response
struct in future to avoid falling into this trap again!